Pictures clicking in Item List View on See Exercises Activity and some bugs fixed
Now when the user click to the item opening the file, the user will see a picture of that exercises. I fixed a bug that I could not choose exercises in Android 4.+ and the Chronometer Activity is a single task, on this way, the chronometer never stop if the user go to another side of application.